moving 'utility' functions/classes out of model_zoo.
