+++ nmt_keras/model_zoo.py
@@ -1338,12 +1338,6 @@ def get_slices3d(x, n):
     dim = int(K.int_shape(x)[2] / n)
     return [Lambda(slice3d, arguments={'dim': dim, 'index': i}, output_shape=lambda s: (s[0], s[1], dim))(x) for i in range(n)]
 
-def merge(x, params, dim):
-    return Lambda(lambda x: K.stack(x,axis=1), output_shape=(params['MAX_OUTPUT_TEXT_LEN'], dim * 2))(x)
-
-def one_hot(x, params):
-    return Lambda(lambda x: K.one_hot(x, params['OUTPUT_VOCABULARY_SIZE']), output_shape=(None, params['OUTPUT_VOCABULARY_SIZE']))(x)
-
 def get_last_state(x, params):
     a = x[:, -1, :]
     return Lambda(lambda x: x[:, -1, :], output_shape=(1, params['QE_VECTOR_SIZE']*2))(x)
@@ -1411,45 +1405,6 @@ class ShiftedConcat(Layer):
         return mask
 
 
-class DenseTranspose(Layer):
-
-  def __init__(self, output_dim, other_layer, other_layer_name, **kwargs):
-      self.output_dim = output_dim
-      self.other_layer=other_layer
-      self.other_layer_name = other_layer_name
-      super(DenseTranspose, self).__init__(**kwargs)
-
-  def call(self, x):
-      # w = self.other_layer.get_layer(self.other_layer_name).layer.kernel
-      w = self.other_layer.layer.kernel
-      w_trans = K.transpose(w)
-      return K.dot(x, w_trans)
-
-  def compute_output_shape(self, input_shape):
-      return (input_shape[0], input_shape[1], self.output_dim)
-
-
-class Reverse(Layer):
-
-    def __init__(self, output_dim, axes, **kwargs):
-        self.output_dim = output_dim
-        self.axes = axes
-        self.supports_masking = True
-        super(Reverse, self).__init__(**kwargs)
-
-    def call(self, x):
-        return K.reverse(x, axes=self.axes)
-
-    def compute_output_shape(self, input_shape):
-        return (input_shape[0],input_shape[1],self.output_dim)
-
-    def compute_mask(self, inputs, mask):
-        if isinstance(mask, list):
-            mask = mask[0]
-        else:
-            mask= None
-        return mask
-
 
 class MyZeroesLayer(Layer):
 
@@ -1470,116 +1425,9 @@ class MyZeroesLayer(Layer):
             mask= None
         return mask
 
-
-class GeneralReshape(Layer):
-
-    def __init__(self, output_dim, params, **kwargs):
-        self.output_dim = output_dim
-        self.params = params
-        super(GeneralReshape, self).__init__(**kwargs)
-
-    def call(self, x):
-        if len(self.output_dim)==2:
-            return K.reshape(x, (-1, self.params['MAX_INPUT_TEXT_LEN']))
-        if len(self.output_dim)==3:
-            return K.reshape(x, (-1, self.output_dim[1], self.output_dim[2]))
-        if len(self.output_dim)==4:
-            return K.reshape(x, (-1, self.output_dim[1], self.output_dim[2], self.output_dim[3]))
-
-    def compute_output_shape(self, input_shape):
-        return self.output_dim
-
-
-def attention_3d_block(inputs, params, ext):
-    '''
-    simple attention: weights over time steps; as in https://github.com/philipperemy/keras-attention-mechanism
-    '''
-    # inputs.shape = (batch_size, time_steps, input_dim)
-    TIME_STEPS = K.int_shape(inputs)[1]
-    input_dim = K.int_shape(inputs)[2]
-
-    a = Permute((2, 1))(inputs)
-    a = Dense(TIME_STEPS, activation='softmax', name='soft_att' + ext)(a)
-    a = Lambda(lambda x: K.mean(x, axis=1), name='dim_reduction' + ext, output_shape=(TIME_STEPS,))(a)
-    a = RepeatVector(input_dim)(a)
-    a_probs = Permute((2, 1), name='attention_vec' + ext)(a)
-
-    output_attention_mul = multiply([inputs, a_probs], name='attention_mul' + ext)
-    sum = Lambda(reduce_sum, mask_aware_mean_output_shape)
-    output = sum(output_attention_mul)
-
-    return output
-
-
-def reduce_max(x):
-    return K.max(x, axis=1, keepdims=False)
-
-
-def reduce_sum(x):
-    return K.sum(x, axis=1, keepdims=False)
-
-
-
-class NonMasking(Layer):
-    def __init__(self, **kwargs):
-        self.supports_masking = True
-        super(NonMasking, self).__init__(**kwargs)
-
-    def build(self, input_shape):
-        input_shape = input_shape
-
-    def compute_mask(self, input, input_mask=None):
-        # do not pass the mask to the next layers
-        return None
-
-    def call(self, x, mask=None):
-        return x
-
-    def compute_output_shape(self, input_shape):
-        return input_shape
-
-
-def mask_aware_mean(x):
-    '''
-    see: https://github.com/keras-team/keras/issues/1579
-    '''
-    # recreate the masks - all zero rows have been masked
-    mask = K.not_equal(K.sum(K.abs(x), axis=2, keepdims=True), 0)
-    # number of that rows are not all zeros
-    n = K.sum(K.cast(mask, 'float32'), axis=1, keepdims=False)
-
-    x_mean = K.sum(x, axis=1, keepdims=False)
-    x_mean = x_mean / n
-
-    return x_mean
-
-
 def mask_aware_mean_output_shape(input_shape):
     shape = list(input_shape)
     assert len(shape) == 3
     return (shape[0], shape[2])
 
 
-def mask_aware_mean4d(x):
-    '''
-    see: https://github.com/keras-team/keras/issues/1579
-    '''
-    # recreate the masks - all zero rows have been masked
-    mask = K.not_equal(K.sum(K.abs(x), axis=3, keepdims=True), 0)
-    # number of that rows are not all zeros
-    n = K.sum(K.cast(mask, 'float32'), axis=2, keepdims=False)
-
-    x_mean = K.sum(x, axis=2, keepdims=False)
-    x_mean = x_mean / n
-
-    return x_mean
-
-def sum4d(x):
-
-    return K.sum(x, axis=2, keepdims=False)
-
-
-def mask_aware_merge_output_shape4d(input_shape):
-    shape = list(input_shape)
-    assert len(shape) == 4
-    return (shape[0], shape[1], shape[3])
